+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Mac builds compiled for x64 processors are now provided in addition to the arm64 builds.
 - demangle: A newline is now printed at the end of the output.
 - stdump: The `functions` and `globals` subcommands now output a header comment like the `types` command.
+- stdump: Added `--compact` option to omit whitespace and newlines from JSON output.
 - stdump: The output format of the `files` subcommand has been updated.
 - stdump: A crash in the command line parser has been fixed.